+++ src/components/owner/OwnerNotices/OwnerNotices.module.css
@@ -2,12 +2,6 @@
   display: flex;
   gap: 4rem;
   flex-direction: column;
-
-  @media (min-width: 768px) {
-  }
-
-  @media (min-width: 1200px) {
-  }
 }
 
 .container {
