+++ src/pages/owner/detail.tsx
@@ -11,8 +11,15 @@ import styles from './detail.module.css';
 import useShopInfo from '@/hooks/useShopInfo';
 import LocationIcon from '@/assets/images/location_icon.svg';
 import OwnerNotices from '@/components/owner/OwnerNotices/OwnerNotices';
+import Pagination from '@/components/@common/pagination/Pagination';
+import getShopsNoticeList from '@/api/getShopsNoticeList';
+import usePagination from '@/hooks/usePagination';
+import { NoticeData } from '@/@types/notice.types';
 
 function Detail() {
+  const [notice, setNotice] = useState<NoticeData>();
+  const { currentPage, totalItems, updateCurrentPage, updateTotalItems } = usePagination({});
+  const { execute: noticeListExecute } = useAsync(getShopsNoticeList);
   const { execute } = useAsync(getShop);
   const { shop } = useOwnerInfo();
   const [hasNotice, setHasNotice] = useState(false);
@@ -61,6 +68,35 @@ function Detail() {
     fetch();
   }, [hasNotice]);
 
+  const noticeListProps = {
+    url: {
+      shopId: id,
+    },
+    params: {
+      offset: (currentPage - 1) * 6,
+      limit: 6,
+    },
+  };
+
+  const handlePageChange = (pageNumber: number) => {
+    updateCurrentPage(pageNumber);
+  };
+
+  const totalPages = Math.ceil(totalItems / 6);
+
+  const noticeListFetch = async () => {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const response: any = await noticeListExecute(noticeListProps);
+    if (response.status === 200) {
+      setNotice(response.data);
+      updateTotalItems(response.data.count);
+    }
+  };
+
+  useEffect(() => {
+    noticeListFetch();
+  }, [currentPage]);
+
   return (
     <>
       <GNBNav userType="employer" />
@@ -90,15 +126,19 @@ function Detail() {
             </div>
           </div>
         </div>
-        {hasNotice ? (
+        {notice ? (
           <div className={styles.secondWrapper}>
             <OwnerNotices
+              notice={notice}
               shopId={id}
               imageUrl={imageUrl}
               name={name}
               address1={address}
               originalHourlyPay={originalHourlyPay}
             />
+            {totalPages > 0 && notice && notice.items.length > 0 && (
+              <Pagination totalPages={totalPages} currentPage={currentPage} onPageChange={handlePageChange} />
+            )}
           </div>
         ) : (
           <div className={styles.secondWrapper}>
